+++ src/main/java/com/community/domain/reply/ReplyRepository.java
@@ -1,5 +1,6 @@
-package com.community.domain.board;
+package com.community.domain.reply;
 
+import com.community.domain.board.Board;
 import com.community.domain.council.Council;
 import com.community.domain.market.Market;
 import com.community.domain.study.Meetings;
