+++ src/main/java/com/community/service/ReportService.java
@@ -2,12 +2,12 @@ package com.community.service;
 
 import com.community.domain.account.Account;
 import com.community.domain.board.Board;
-import com.community.domain.board.Reply;
+import com.community.domain.reply.Reply;
 import com.community.domain.report.BoardReport;
 import com.community.domain.report.ReplyReport;
 import com.community.web.dto.BoardReportForm;
 import com.community.domain.board.BoardRepository;
-import com.community.domain.board.ReplyRepository;
+import com.community.domain.reply.ReplyRepository;
 import com.community.web.dto.ReplyReportForm;
 import com.community.domain.report.BoardReportRepository;
 import com.community.domain.report.ReplyReportRepository;
