+++ src/main/java/com/community/web/controller/BoardController.java
@@ -4,20 +4,16 @@ import com.community.domain.account.Account;
 import com.community.domain.account.AccountType;
 import com.community.domain.account.CurrentUser;
 import com.community.domain.board.Board;
-import com.community.domain.board.Reply;
-import com.community.domain.bookmark.Bookmark;
 import com.community.domain.bookmark.BookmarkRepository;
 import com.community.infra.config.SecurityUser;
 import com.community.web.dto.BoardForm;
 import com.community.web.dto.ReplyForm;
 import com.community.web.dto.BoardReportForm;
 import com.community.domain.board.BoardRepository;
-import com.community.domain.board.ReplyRepository;
+import com.community.domain.reply.ReplyRepository;
 import com.community.service.BoardService;
 import com.community.service.ReplyService;
 import com.community.domain.likes.LikeRepository;
-import com.community.domain.likes.Likes;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
