+++ src/main/java/com/community/web/controller/MainController.java
@@ -3,9 +3,9 @@ package com.community.web.controller;
 import com.community.domain.account.Account;
 import com.community.domain.account.CurrentUser;
 import com.community.domain.board.Board;
-import com.community.domain.board.Reply;
+import com.community.domain.reply.Reply;
 import com.community.domain.board.BoardRepository;
-import com.community.domain.board.ReplyRepository;
+import com.community.domain.reply.ReplyRepository;
 import com.community.domain.study.StudyRepository;
 import com.community.service.BoardService;
 import com.community.service.ReplyService;
