+++ src/main/java/com/community/web/controller/ManageController.java
@@ -4,8 +4,8 @@ import com.community.domain.account.Account;
 import com.community.domain.account.CurrentUser;
 import com.community.domain.board.Board;
 import com.community.domain.board.BoardRepository;
-import com.community.domain.board.Reply;
-import com.community.domain.board.ReplyRepository;
+import com.community.domain.reply.Reply;
+import com.community.domain.reply.ReplyRepository;
 import com.community.domain.inquire.Inquire;
 import com.community.domain.inquire.InquireRepository;
 import com.community.domain.report.BoardReport;
