+++ src/main/java/com/community/web/controller/ReplyController.java
@@ -7,9 +7,9 @@ import com.community.domain.market.Market;
 import com.community.domain.market.MarketRepository;
 import com.community.infra.alarm.ReplyCreatePublish;
 import com.community.domain.board.Board;
-import com.community.domain.board.Reply;
+import com.community.domain.reply.Reply;
 import com.community.domain.board.BoardRepository;
-import com.community.domain.board.ReplyRepository;
+import com.community.domain.reply.ReplyRepository;
 import com.community.service.ReplyService;
 import com.community.domain.council.Council;
 import com.community.domain.council.CouncilRepository;
