+++ src/main/java/com/community/web/controller/ReportController.java
@@ -3,9 +3,9 @@ package com.community.web.controller;
 import com.community.domain.account.CurrentUser;
 import com.community.domain.account.Account;
 import com.community.domain.board.Board;
-import com.community.domain.board.Reply;
+import com.community.domain.reply.Reply;
 import com.community.domain.board.BoardRepository;
-import com.community.domain.board.ReplyRepository;
+import com.community.domain.reply.ReplyRepository;
 import com.community.domain.council.Council;
 import com.community.domain.council.CouncilRepository;
 import com.community.domain.market.Market;
