+++ src/main/java/com/community/web/controller/StudyController.java
@@ -9,10 +9,10 @@ import com.community.domain.study.StudyRepository;
 import com.community.infra.alarm.MeetingCreatedPublish;
 import com.community.infra.alarm.ReplyCreatePublish;
 import com.community.infra.alarm.StudyCreatedPublish;
-import com.community.domain.board.Reply;
+import com.community.domain.reply.Reply;
 import com.community.infra.aws.S3Service;
 import com.community.web.dto.ReplyForm;
-import com.community.domain.board.ReplyRepository;
+import com.community.domain.reply.ReplyRepository;
 import com.community.service.ReplyService;
 import com.community.service.StudyService;
 import com.community.domain.study.Meetings;
